+++ app/(dashboard)/finances/page.tsx
@@ -10,30 +10,7 @@ import Container from '@/app/components/Container';
 import EmbeddedComponentContainer from '@/app/components/EmbeddedComponentContainer';
 import {LandmarkIcon, LoaderCircle} from 'lucide-react';
 import {Button} from '@/components/ui/button';
-
-export const useFinancialAccount = () => {
-  const [financialAccount, setFinancialAccount] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        const response = await fetch('/api/financial_account');
-        const json = await response.json();
-        setFinancialAccount(json.financial_account);
-        setLoading(false);
-      } catch (error: any) {
-        setError(error);
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, []);
-
-  return {loading, financialAccount, error};
-};
+import {useFinancialAccount} from '@/app/hooks/useFinancialAccount';
 
 export default function Finances() {
   const {
