--- app/components/CurrencyInput.tsx
@@ -1,60 +1,0 @@
-import React from 'react';
-
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import {ControllerRenderProps} from 'react-hook-form';
-
-type Props = {
-  value: unknown;
-  onChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
-};
-
-type Currency =
-  | 'aed'
-  | 'aud'
-  | 'cad'
-  | 'cny'
-  | 'eur'
-  | 'gbp'
-  | 'inr'
-  | 'jpy'
-  | 'sgd'
-  | 'usd';
-const CurrencyOptions: Record<Currency, string> = {
-  aed: 'AED',
-  aud: 'AUD',
-  cad: 'CAD',
-  cny: 'CNY',
-  eur: 'EUR',
-  gbp: 'GBP',
-  inr: 'INR',
-  jpy: 'JPY',
-  sgd: 'SGD',
-  usd: 'USD',
-};
-
-function CurrencySelect({
-  field,
-}: {
-  field: ControllerRenderProps<z.infer<typeof formSchema>, 'currency'>;
-}) {
-  return (
-    <Select {...field} onValueChange={(value) => field.onChange(value)}>
-      <SelectTrigger className="mt-1">
-        <SelectValue>{CurrencyOptions[field.value as Currency]}</SelectValue>
-      </SelectTrigger>
-      <SelectContent>
-        {Object.keys(CurrencyOptions).map((key) => (
-          <SelectItem key={key} value={CurrencyOptions[key as Currency]}>
-            {CurrencyOptions[key as Currency]}
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  );
-}
