+++ app/hooks/useFinancialAccount.ts
@@ -1,0 +1,25 @@
+import {useEffect, useState} from 'react';
+
+export const useFinancialAccount = () => {
+  const [financialAccount, setFinancialAccount] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      setLoading(true);
+      try {
+        const response = await fetch('/api/financial_account');
+        const json = await response.json();
+        setFinancialAccount(json.financial_account);
+        setLoading(false);
+      } catch (error: any) {
+        setError(error);
+        setLoading(false);
+      }
+    };
+    fetchData();
+  }, []);
+
+  return {loading, financialAccount, error};
+};
