+++ package.json
@@ -44,7 +44,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.51.1",
-    "sharp": "^0.33.3",
+    "sharp": "^0.33.4",
     "stripe": "^15.4.0-beta.1",
     "tailwind-merge": "^2.2.2",
     "tailwindcss-animate": "^1.0.7",
